image-service/requirements.txt
fastapi
uvicorn[standard]
tritonclient
torchvision
grpcio-tools
python-multipart